SUI: Implemented fragments for Vigenere. Add images to fragment in future.
