+++ app/src/main/java/com/example/cipherhub/CaesarActivity.java
@@ -1,17 +1,13 @@
 package com.example.cipherhub;
 
 import android.os.Bundle;
-import android.text.Editable;
-import android.text.TextWatcher;
-import android.widget.EditText;
 
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.viewpager.widget.ViewPager;
 
-import ciphers.CaesarCipher;
 import ui.CipherFragment;
 import ui.SectionFragment;
-import ui.SectionPageAdapter;
+import ui.FragmentPageAdapter;
 import ui.SetUpPagerInterface;
 
 public class CaesarActivity extends AppCompatActivity implements SetUpPagerInterface {
@@ -19,10 +15,10 @@ public class CaesarActivity extends AppCompatActivity implements SetUpPagerInter
     //include viewpager for activities -> done
 
     ViewPager viewPager;
-    SectionPageAdapter sectionPageAdapter;
+    FragmentPageAdapter fragmentPageAdapter;
 
     public void setUpViewPager(ViewPager viewPager) {
-        SectionPageAdapter adapter = new SectionPageAdapter(getSupportFragmentManager());
+        FragmentPageAdapter adapter = new FragmentPageAdapter(getSupportFragmentManager());
         adapter.addFragment(new SectionFragment(), "Caesar Screen");
         adapter.addFragment(new CipherFragment(), "Caesar Cipher");
         viewPager.setAdapter(adapter);
@@ -34,7 +30,7 @@ public class CaesarActivity extends AppCompatActivity implements SetUpPagerInter
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_caesar);
 
-        sectionPageAdapter = new SectionPageAdapter(getSupportFragmentManager());
+        fragmentPageAdapter = new FragmentPageAdapter(getSupportFragmentManager());
         viewPager = findViewById(R.id.container);
 
         setUpViewPager(viewPager);
