+++ app/src/main/res/layout/activity_caesar.xml
@@ -11,16 +11,4 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
 
-    <TextView
-        android:id="@+id/CaesarOutput"
-        android:layout_width="110dp"
-        android:layout_height="36dp"
-        android:layout_marginStart="176dp"
-        android:layout_marginLeft="176dp"
-        android:layout_marginEnd="176dp"
-        android:layout_marginRight="176dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        tools:layout_editor_absoluteY="214dp" />
-
 </androidx.constraintlayout.widget.ConstraintLayout>
